[ALPS06391680] mml: dma_buf_map refactory
1.Disable m6893/mt6895/mt6879 build for mml
2.Fix dma_buf_map vaddr type error

MTK-Commit-Id: 9e06c4b926feba44cefb301016cbb152fadaba24

CR-Id: ALPS06391680
Feature: [Module]MDP Driver

+++ drivers/gpu/drm/mediatek/mml/mtk-mml-buf.c
@@ -120,7 +120,7 @@ int mml_buf_va_get(struct mml_file_buf *buf)
 		ret = dma_buf_vmap(buf->dma[i].dmabuf, &map);
 		if (ret)
 			return -ENOMEM;
-		buf->dma[i].va = (uint64_t)map.vaddr;
+		buf->dma[i].va = map.vaddr;
 	}
 
 	return 0;
